Add frame stack and keys to edit it
Add a stack that can store frames to be used over and over again. Add
three keys to alter the stack:

* key 'd': push last frame of movie onto stack

* key 's': set last frame of movie to frame on top of stack (no poping)

* key 'a': remove frame on top of stack

The animator is too time-consuming and dificult to use. Objects in the
background and other objects that aren't moving must be drawn over and
over again making the animator much less entertaining.

* Declare key variables for 'd', 's', and 'a'; and declare stack

* Make event listener update stack-control key variables

* Make animator do things when stack-control keys are pressed